cap_agent/pyproject.toml
[project]
name = "cap-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.22",
    "google-adk>=1.23.0",
    "httpx>=0.28.1",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.2.1",
]
